requirements.txt
aiogram
selenium
python-dotenv